Full codebase conversion to TypeScript
