+++ .github/workflows/ci-module.yml
@@ -4,11 +4,10 @@ on:
   push:
     branches:
       - master
+      - next
   pull_request:
   workflow_dispatch:
 
 jobs:
   test:
-    uses: hapijs/.github/.github/workflows/ci-module.yml@master
-    with:
-      min-node-version: 14
+    uses: hapijs/.github/.github/workflows/ci-module.yml@min-node-20-hapi-21
