+++ package.json
@@ -3,10 +3,11 @@
   "description": "Mime database interface",
   "version": "7.0.1",
   "repository": "git://github.com/hapijs/mimos",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
+  "main": "dist/index.js",
+  "type": "module",
+  "types": "dist/index.d.ts",
   "files": [
-    "lib"
+    "dist"
   ],
   "keywords": [
     "mime",
@@ -16,6 +17,16 @@
   "eslintConfig": {
     "extends": [
       "plugin:@hapi/module"
+    ],
+    "parserOptions": {
+      "sourceType": "module",
+      "ecmaVersion": 2022
+    },
+    "overrides": [
+      {
+        "files": ["*.ts"],
+        "parser": "@typescript-eslint/parser"
+      }
     ]
   },
   "dependencies": {
@@ -25,13 +36,17 @@
   "devDependencies": {
     "@hapi/code": "^9.0.0",
     "@hapi/eslint-plugin": "^6.0.0",
-    "@hapi/lab": "^25.0.1",
-    "@types/node": "^17.0.31",
-    "typescript": "~4.6.4"
+    "@hapi/lab": "^25.0.0",
+    "@types/node": "^20",
+    "@typescript-eslint/parser": "^7.0.0",
+    "tsx": "^4.19.0",
+    "typescript": "^5.8"
   },
   "scripts": {
-    "test": "lab -m 5000 -t 100 -L -a @hapi/code -Y",
-    "test-cov-html": "lab -a @hapi/code -r html -o coverage.html -L"
+    "test": "node --import tsx --test --experimental-test-coverage --test-coverage-branches=100 --test-coverage-functions=100 --test-coverage-lines=100 --test-coverage-exclude=test/** test/index.ts",
+    "lint": "tsc --noEmit && eslint . --ext .js,.ts",
+    "build": "tsc -p tsconfig.build.json",
+    "prepublishOnly": "npm run build"
   },
   "license": "BSD-3-Clause"
 }
