+++ tsconfig.build.json
@@ -1,0 +1,10 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "noEmit": false,
+        "outDir": "dist"
+    },
+    "include": [
+        "lib/**/*.ts"
+    ]
+}
